WebCam_YOLO/pyproject.toml
[project]
name = "thomas-ai-activite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.7.0",
    "diffusers>=0.33.1",
    "lap>=0.5.12",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "ultralytics>=8.3.139",
]
